packages/stargate-agent/pyproject.toml
[project]
name = "stargate-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.26",
    "mcp[server]>=1.5.0",
    "openai>=1.30.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "mcp-ui-server>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stargate_agent"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
]
